unable save and delete for models with is_serializable set to false

+++ src/Comhon/Serialization/ValidatedSerializationUnit.php
@@ -18,8 +18,11 @@ use Comhon\Exception\ArgumentException;
 abstract class ValidatedSerializationUnit extends SerializationUnit {
 	
 	public final function validateSerialization(UniqueObject $object) {
+		if (!$object->getModel()->hasSerialization()) {
+			throw new SerializationException("object with model '{$object->getModel()->getName()}' doesn't have serialization");
+		}
 		if (is_null($object->getModel()->getSerialization()->getSerializationUnit())) {
-			throw new SerializationException("object with model '{$object->getModel()}' doesn't have serialization unit");
+			throw new SerializationException("object with model '{$object->getModel()->getName()}' doesn't have serialization unit");
 		}
 		if (!is_null($object->getModel()->getSerializationSettings())) {
 			if ($object->getModel()->getSerializationSettings()->getModel()->getName() !== static::getType()) {
@@ -46,7 +49,9 @@ abstract class ValidatedSerializationUnit extends SerializationUnit {
 	 */
 	public final function saveObject(UniqueObject $object, $operation = null) {
 		$this->validateSerialization($object);
-		
+		if (!$object->getModel()->getSerialization()->isSerializationAllowed()) {
+			throw new SerializationException("object with model '{$object->getModel()->getName()}' doesn't have serialization allowed");
+		}
 		if (!is_null($operation) && ($operation !== self::CREATE) && ($operation !== self::UPDATE)) {
 			throw new ArgumentException($operation, [self::CREATE, self::UPDATE], 2);
 		}
@@ -72,6 +77,9 @@ abstract class ValidatedSerializationUnit extends SerializationUnit {
 	 */
 	public final function deleteObject(UniqueObject $object) {
 		$this->validateSerialization($object);
+		if (!$object->getModel()->getSerialization()->isSerializationAllowed()) {
+			throw new SerializationException("object with model '{$object->getModel()->getName()}' doesn't have serialization allowed");
+		}
 		return $this->_deleteObject($object);
 	}
